create id table based on wikidata
In order to map different ids to each other, i combine wikidata queries to extract ids from deezer, mbz, discogs etc and i create a table with all the matches.

Then i incorporate mbz-deezer pairings, which adds a lot of deezer_ids which were missing before.

Now one challenge is to reduce the amount of duplicates: for one deezer_id, there are sometimes dozens of mbz or spotify identifiers.

Another challenge is to retrieve the wikidata labels in one query, which is somehow difficult; i left it out for now.