consolidate id table
full-joining all the different ids together leads to the combinatory issue that the number of cases explodes exponentially. so first leave spotify and discogs out to focus on the main identifiers, and perform some sanity checks

also, there seems to be an issue with contacts.csv, which leads to a super high number of weird duplicates. need to further check that.